First test run
There a multitude of changes made here. Most importantly, the interrupts were changed to callbacks. This then resulted in the use of 6 GPIO pins instead of 3. That's because there cannot be two separate callbacks for 1 GPIO line that would detect a rising or falling edge. 
The countdown variables were  commented out because there was a global declaration issue.